Made Range, RangeSet, and Session all use proper RFC1982 comparisons per QPID-861. Also switched command ids from long -> int, and added a mutex to channel to prevent multi-frame commands from interleaving when invoked from separate threads.
git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/qpid/trunk/qpid@659271 13f79535-47bb-0310-9956-ffa450edef68

+++ java/client/src/main/java/org/apache/qpid/client/BasicMessageConsumer_0_10.java
@@ -358,7 +358,7 @@ public class BasicMessageConsumer_0_10 extends BasicMessageConsumer<Struct[], By
         if (!_preAcquire)
         {
             RangeSet ranges = new RangeSet();
-            ranges.add(message.getDeliveryTag());
+            ranges.add((int) message.getDeliveryTag());
             _0_10session.getQpidSession().messageAcknowledge(ranges);
             _0_10session.getCurrentException();
         }
@@ -375,7 +375,7 @@ public class BasicMessageConsumer_0_10 extends BasicMessageConsumer<Struct[], By
         if (_preAcquire)
         {
             RangeSet ranges = new RangeSet();
-            ranges.add(message.getDeliveryTag());
+            ranges.add((int) message.getDeliveryTag());
             _0_10session.getQpidSession().messageRelease(ranges);
             _0_10session.getCurrentException();
         }
@@ -394,7 +394,7 @@ public class BasicMessageConsumer_0_10 extends BasicMessageConsumer<Struct[], By
         if (!_preAcquire)
         {
             RangeSet ranges = new RangeSet();
-            ranges.add(message.getDeliveryTag());
+            ranges.add((int) message.getDeliveryTag());
 
             Acquired acq = _0_10session.getQpidSession().messageAcquire(ranges).get();
