+++ java/client/src/main/java/org/apache/qpidity/nclient/util/StreamingMessage.java
@@ -61,7 +61,7 @@ public class StreamingMessage extends ReadOnlyMessage implements Message
      * does not have a transfer id. Hence this method is not
      * applicable to this implementation.    
      */
-    public long getMessageTransferId()
+    public int getMessageTransferId()
     {
         throw new UnsupportedOperationException();
     }
